+++ Makefile
@@ -19,7 +19,7 @@ SRCS = cmd_cmd_invocation.c cmd_cmd_invocation2.c cmd_exec_command.c		\
 	builtin.c builtin_echo.c builtin_env.c builtin_exit.c					\
 	builtin_cd.c builtin_cd_path.c builtin_cd_chdir.c builtin_cd_cdpath.c	\
 	builtin_export.c builtin_pwd.c builtin_unset.c							\
-	str_utils.c shell_initialization.c
+	str_utils.c shell_initialization.c cmd_exec_builtin.c
 
 OBJS = $(SRCS:.c=.o)
