+++ cmd_redirection.c
@@ -8,47 +8,6 @@
 #include "env.h"
 #include "minishell.h"
 
-t_fd_list	*fd_list_add_fd(t_fd_list **lst, int fd)
-{
-	t_fd_list	*new_lst;
-	t_fd_list	*tmp;
-
-	if (!lst)
-		return (NULL);
-	new_lst = ft_calloc(1, sizeof(t_fd_list));
-	new_lst->fd = fd;
-	if (!new_lst)
-		return (NULL);
-	if (!*lst)
-		*lst = new_lst;
-	else
-	{
-		tmp = *lst;
-		while (tmp->next)
-			tmp = tmp->next;
-		tmp->next = new_lst;
-	}
-	return (new_lst);
-}
-
-void		fd_list_close(t_fd_list **lst)
-{
-	t_fd_list	*current;
-	t_fd_list	*tmp;
-
-	if (!lst)
-		return ;
-	current = *lst;
-	while (current)
-	{
-		close(current->fd);
-		tmp = current;
-		current = current->next;
-		free(tmp);
-	}
-	*lst = NULL;
-}
-
 /*
  * リダイレクション用に渡された文字列の環境変数展開を行う.
  *
